refactor: Make results::action_result private
