+++ src/results/mod.rs
@@ -1,3 +1,5 @@
-pub mod action_result;
+pub use action_result::{ActionResult, CreateFileResult, ScriptResult, VerifyResult};
+
+mod action_result;
 pub mod basic_printer;
 pub mod printer;
